Moving stub implementation code to another file, getgauge/gauge-vscode#156
