+++ api/lang/customResponses.go
@@ -22,7 +22,6 @@ import (
 	"fmt"
 
 	"github.com/getgauge/gauge/gauge"
-	gm "github.com/getgauge/gauge/gauge_messages"
 	"github.com/getgauge/gauge/parser"
 	"github.com/getgauge/gauge/util"
 	"github.com/sourcegraph/go-langserver/pkg/lsp"
@@ -40,11 +39,6 @@ type specInfo struct {
 	ExecutionIdentifier string `json:"executionIdentifier"`
 }
 
-type stubImpl struct {
-	ImplementationFilePath string   `json:"implementationFilePath"`
-	Codes                  []string `json:"codes"`
-}
-
 func specs() (interface{}, error) {
 	specDetails := provider.GetAvailableSpecDetails([]string{})
 	specs := make([]specInfo, 0)
@@ -52,55 +46,6 @@ func specs() (interface{}, error) {
 		specs = append(specs, specInfo{Heading: d.Spec.Heading.Value, ExecutionIdentifier: d.Spec.FileName})
 	}
 	return specs, nil
-}
-
-func getImplFiles(req *jsonrpc2.Request) (interface{}, error) {
-	var info = struct {
-		Concept bool `json:"concept"`
-	}{}
-	if err := json.Unmarshal(*req.Params, &info); err != nil {
-		return nil, fmt.Errorf("failed to parse request %s", err.Error())
-	}
-	if info.Concept {
-		return util.GetConceptFiles(), nil
-	}
-	implementationFileListResponse, err := getImplementationFileList()
-	if err != nil {
-		return nil, err
-	}
-	return implementationFileListResponse.ImplementationFilePaths, nil
-}
-
-func putStubImpl(req *jsonrpc2.Request) (interface{}, error) {
-	var stubImplParams stubImpl
-	if err := json.Unmarshal(*req.Params, &stubImplParams); err != nil {
-		return nil, fmt.Errorf("failed to parse request %s", err)
-	}
-	fileDiff, err := putStubImplementation(stubImplParams.ImplementationFilePath, stubImplParams.Codes)
-	if err != nil {
-		return nil, err
-	}
-
-	return getWorkspaceEditForStubImpl(fileDiff), nil
-}
-
-func getWorkspaceEditForStubImpl(fileDiff *gm.FileDiff) lsp.WorkspaceEdit {
-	var result lsp.WorkspaceEdit
-	result.Changes = make(map[string][]lsp.TextEdit, 0)
-	uri := util.ConvertPathToURI(lsp.DocumentURI(fileDiff.FilePath))
-
-	var textDiffs = fileDiff.TextDiffs
-	for _, textDiff := range textDiffs {
-		textEdit := lsp.TextEdit{
-			NewText: textDiff.Content,
-			Range: lsp.Range{
-				Start: lsp.Position{Line: int(textDiff.Span.Start), Character: int(textDiff.Span.StartChar)},
-				End:   lsp.Position{Line: int(textDiff.Span.End), Character: int(textDiff.Span.EndChar)},
-			},
-		}
-		result.Changes[string(uri)] = append(result.Changes[string(uri)], textEdit)
-	}
-	return result
 }
 
 func scenarios(req *jsonrpc2.Request) (interface{}, error) {
